Using vision yaw with GPS and without mag-heading initialization
This is a dirty, but working version, that makes it possible to fuse a vision
yaw even if GPS fusion is enabled. Also, we doesn't initiate EKF until a first
ev_pos message is received. The initiation is changed regarding to vision yaw,
because in current version in 1.8.2 release, the heading initialed only by
a magnetometer value at first. It's unacceptable behaviour in cases, when we are
starting in the enclosures, that has a lot of the metal or magnet things that
interfere with a drone's magnetometer.

This code has been tested many times on the real flights and it didn't show any
visible troubles.

It also related to: https://github.com/PX4/ecl/issues/397